chore: move tipo para arquivo correto
