--- src/app_types.ts
@@ -1,6 +1,0 @@
-type OptionalCourse = {
-  label: string
-  workloadInHours: number
-}
-
-export type { OptionalCourse }
